[TEST] - Added tests for FileDependency
